refactor(create-a-container/client): move MultiSelect into it's own file
